Interactable and Pickupable Working
I improved the interactability of the game.  Not objects in the game can either be pickupable or interactable.  If they are pickupable the player can pick them up and if they are interactable the player can interact with them if they are holding a correct item.  At the moment if the player is holding a item and they walk over to another item after a certain time the second item will be destroyed.  This only works for the Mop and Sphere objects in the game currently.